merge latest master

+++ src/sentry/api/endpoints/project_search_details.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from rest_framework import serializers
 from rest_framework.response import Response
 
-from sentry.api.bases.project import ProjectEndpoint, ProjectPermission
+from sentry.api.bases.project import ProjectEndpoint, RelaxedSearchPermission
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.api.serializers import serialize
 from sentry.models import SavedSearch, SavedSearchUserDefault
@@ -20,16 +20,6 @@ class SavedSearchSerializer(serializers.Serializer):
     isUserDefault = serializers.BooleanField(required=False)
 
 
-class RelaxedSearchPermission(ProjectPermission):
-    scope_map = {
-        'GET': ['project:read', 'project:write', 'project:delete'],
-        'POST': ['project:write', 'project:delete'],
-        # members can do partial writes
-        'PUT': ['project:write', 'project:delete', 'project:read'],
-        'DELETE': ['project:delete'],
-    }
-
-
 class ProjectSearchDetailsEndpoint(ProjectEndpoint):
     permission_classes = (RelaxedSearchPermission,)
 
@@ -43,10 +33,7 @@ class ProjectSearchDetailsEndpoint(ProjectEndpoint):
 
         """
         try:
-            search = SavedSearch.objects.get(
-                project=project,
-                id=search_id,
-            )
+            search = SavedSearch.objects.get(project=project, id=search_id)
         except SavedSearch.DoesNotExist:
             raise ResourceDoesNotExist
 
@@ -67,46 +54,42 @@ class ProjectSearchDetailsEndpoint(ProjectEndpoint):
 
         """
         try:
-            search = SavedSearch.objects.get(
-                project=project,
-                id=search_id,
-            )
+            search = SavedSearch.objects.get(project=project, id=search_id)
         except SavedSearch.DoesNotExist:
             raise ResourceDoesNotExist
 
-        if request.access.has_team_scope(project.team, 'project:write'):
-            serializer = SavedSearchSerializer(data=request.DATA, partial=True)
+        has_team_scope = any(
+            request.access.has_team_scope(team, "project:write") for team in project.teams.all()
+        )
+        if has_team_scope:
+            serializer = SavedSearchSerializer(data=request.data, partial=True)
         else:
-            serializer = LimitedSavedSearchSerializer(data=request.DATA, partial=True)
+            serializer = LimitedSavedSearchSerializer(data=request.data, partial=True)
 
         if not serializer.is_valid():
             return Response(serializer.errors, status=400)
 
-        result = serializer.object
+        result = serializer.validated_data
 
         kwargs = {}
-        if result.get('name'):
-            kwargs['name'] = result['name']
-        if result.get('query'):
-            kwargs['query'] = result['query']
-        if result.get('isDefault'):
-            kwargs['is_default'] = result['isDefault']
+        if result.get("name"):
+            kwargs["name"] = result["name"]
+        if result.get("query"):
+            kwargs["query"] = result["query"]
+        if result.get("isDefault"):
+            kwargs["is_default"] = result["isDefault"]
 
         if kwargs:
             search.update(**kwargs)
 
-        if result.get('isDefault'):
-            SavedSearch.objects.filter(
-                project=project,
-            ).exclude(id=search_id).update(is_default=False)
+        if result.get("isDefault"):
+            SavedSearch.objects.filter(project=project).exclude(id=search_id).update(
+                is_default=False
+            )
 
-        if result.get('isUserDefault'):
+        if result.get("isUserDefault"):
             SavedSearchUserDefault.objects.create_or_update(
-                user=request.user,
-                project=project,
-                values={
-                    'savedsearch': search,
-                }
+                user=request.user, project=project, values={"savedsearch": search}
             )
 
         return Response(serialize(search, request.user))
@@ -121,13 +104,18 @@ class ProjectSearchDetailsEndpoint(ProjectEndpoint):
 
         """
         try:
-            search = SavedSearch.objects.get(
-                project=project,
-                id=search_id,
-            )
+            search = SavedSearch.objects.get(project=project, id=search_id)
         except SavedSearch.DoesNotExist:
             raise ResourceDoesNotExist
 
-        search.delete()
+        is_search_owner = request.user and request.user == search.owner
 
-        return Response(status=204)
+        if request.access.has_scope("project:write"):
+            if not search.owner or is_search_owner:
+                search.delete()
+                return Response(status=204)
+        elif is_search_owner:
+            search.delete()
+            return Response(status=204)
+
+        return Response(status=403)
